+++ cmake/consumer-test/main.cpp
@@ -7,8 +7,13 @@
 #endif
 
 int main() {
+  using skymizer::minifloat::BF;
   using skymizer::minifloat::E4M3;
 
   const E4M3 value{1.5F};
-  return value.to_float() == 1.5F && std::numeric_limits<E4M3>::is_specialized ? 0 : 1;
+  const BF<20> wide{BF<16>{1.5F}};
+  return value.to_float() == 1.5F && std::numeric_limits<E4M3>::is_specialized &&
+                 (wide * BF<20>{0.5F}).to_float() == 0.75F
+             ? 0
+             : 1;
 }